Build AnimationSwitcher at one place
Building different AnimationSwitcher for a fade and a flip animation doesn't seem to make sense because they have to have the same duration, switchInCurve and switchOutCurve.

If we specify different values for these properties, the dismissing animation will use the old values while the presenting animation will use the new values, which results in out-of-sync animations.